refactor(settings): extract Mobile / Web Push section to _components/
PR-3 of the settings-page split. Pulls the entire Mobile section
(#s-mobile) into a self-contained MobileSection component.

Moved into the child: pushStatus / pushBusy / pushMsg / haltsPref
state, the SW-status + halt-prefs-load effects, and all four
handlers (subscribe, unsubscribe, test, halt-prefs toggle). The
section talks only to /api/push/* + the service worker — no shared
`settings` slice — so the only parent dependency is `flashSaved`.

Also extracts the shared inline-style objects (inputStyle,
labelStyle, helpStyle) into _components/styles.ts so every
extracted section can import them rather than reaching back into
page.tsx. The push-lib imports, now unused in page.tsx, are
dropped.

The test-POST AbortController, previously the page-level
`abortRef`, is now owned by the section.

Behavior identical. page.tsx: 1710 → 1300 lines.
